requirements.txt
scapy==2.5.0
python-nmap==0.7.1
colorama==0.4.6
netfilterqueue==1.0.0
requests==2.31.0
pyfiglet==1.0.2 